+++ src/ShaderPlayground.Core/CompilerNames.cs
@@ -22,5 +22,6 @@
         public const string ZstdCompress = "zstd-compress";
         public const string ZstdDecompress = "zstd-decompress";
         public const string MinizCompress = "miniz-compress";
+        public const string LzmaCompress = "lzma-compress";
     }
 }