+++ src/ShaderPlayground.Core/LanguageNames.cs
@@ -7,6 +7,7 @@
         public const string SpirV = "SPIR-V";
         public const string Dxbc = "DXBC";
         public const string Dxil = "DXIL";
+        public const string Lzma = "lzma";
         public const string Metal = "Metal";
         public const string Cpp = "C++";
         public const string Ispc = "ISPC";